chore: refactor components into own files
